requirements.txt
cupy-cuda102==7.6.0
Cython==0.29.21
fastrlock==0.5
h5py==2.10.0
numexpr==2.7.1
numpy==1.19.0
opencv-python==4.3.0.36
six==1.15.0
tqdm==4.48.0
